+++ src/Component/OpenApi2/Tests/fixtures/docker-api/expected/Validator/MountVolumeOptionsValidator.php
@@ -6,7 +6,7 @@ class MountVolumeOptionsValidator implements \Docker\Api\Validator\ValidatorInte
 {
     public function validate($data) : void
     {
-        $constraints = array(new \Symfony\Component\Validator\Constraints\Collection(array('fields' => array('NoCopy' => new \Symfony\Component\Validator\Constraints\Optional(array(new \Symfony\Component\Validator\Constraints\Type(array('0' => 'bool'))))), 'allowExtraFields' => true)));
+        $constraints = array(new \Symfony\Component\Validator\Constraints\Collection(array('fields' => array('NoCopy' => new \Symfony\Component\Validator\Constraints\Optional(array(new \Symfony\Component\Validator\Constraints\Type(array('0' => 'bool')))), 'Labels' => new \Symfony\Component\Validator\Constraints\Optional(array()), 'DriverConfig' => new \Symfony\Component\Validator\Constraints\Optional(array())), 'allowExtraFields' => true)));
         $validator = \Symfony\Component\Validator\Validation::createValidator();
         $violations = $validator->validate($data, $constraints);
         if ($violations->count() > 0) {