+++ src/Component/OpenApi2/Tests/fixtures/docker-api/expected/Validator/NetworksIdConnectPostBodyValidator.php
@@ -6,7 +6,7 @@ class NetworksIdConnectPostBodyValidator implements \Docker\Api\Validator\Valida
 {
     public function validate($data) : void
     {
-        $constraints = array(new \Symfony\Component\Validator\Constraints\Collection(array('fields' => array('Container' => new \Symfony\Component\Validator\Constraints\Optional(array(new \Symfony\Component\Validator\Constraints\Type(array('0' => 'string'))))), 'allowExtraFields' => true)));
+        $constraints = array(new \Symfony\Component\Validator\Constraints\Collection(array('fields' => array('Container' => new \Symfony\Component\Validator\Constraints\Optional(array(new \Symfony\Component\Validator\Constraints\Type(array('0' => 'string')))), 'EndpointConfig' => new \Symfony\Component\Validator\Constraints\Optional(array())), 'allowExtraFields' => true)));
         $validator = \Symfony\Component\Validator\Validation::createValidator();
         $violations = $validator->validate($data, $constraints);
         if ($violations->count() > 0) {