+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- [JsonSchema] [GH#629](https://github.com/janephp/janephp/pull/629) Handle validation when fields has no validation guess
+
 ## [7.2.5] - 2022-07-01
 ### Added
 - [JsonSchema] [GH#627](https://github.com/janephp/janephp/pull/627) Added a way to skip validation in Normalizer context